+++ components/cms-modern/ContentBlock/ContentBlock.tsx
@@ -14,9 +14,9 @@ import CmsContentItem from '@components/cms-modern/CmsContentItem';
 import CmsEdition from '@components/cms-modern/CmsEdition';
 import CmsSlot from '@components/cms-modern/CmsSlot';
 import Container from '@components/cms-modern/Container';
-import CuratedProductGrid from '@components/cms/CuratedProductGrid';
+import CuratedProductGrid from '@components/cms-modern/CuratedProductGrid';
 import CustomRichText from '@components/cms-modern/CustomRichText';
-import DynamicBlogList from '@components/cms/DynamicBlogList';
+import DynamicBlogList from '@components/cms-modern/DynamicBlogList';
 import ExternalBlock from '@components/cms-modern/ExternalBlock';
 import FlexibleSlot from '@components/cms-modern/FlexibleSlot';
 import Image from '@components/cms-modern/Image';
