+++ components/cms-modern/CuratedProductGrid/CuratedProductGrid.tsx
