+++ components/cms-modern/CuratedProductGrid/CuratedProductGridCard/CuratedProductGridCard.tsx
