+++ components/cms-modern/CuratedProductGrid/CuratedProductGridCard/index.ts
