+++ components/cms-modern/CuratedProductGrid/index.ts
