+++ components/cms-modern/DynamicBlogList/DynamicBlogListCard.tsx
