+++ components/cms/index.ts
@@ -1,5 +1,5 @@
 export { default as BlogSnippet } from './BlogSnippet';
 export { default as ProductGrid } from './ProductGrid';
-export { default as CuratedProductGrid } from './CuratedProductGrid';
+export { default as CuratedProductGrid } from '../cms-modern/CuratedProductGrid';
 export { default as BlogList } from './BlogList';
-export { default as DynamicBlogList } from './DynamicBlogList';
+export { default as DynamicBlogList } from '../cms-modern/DynamicBlogList';
